backend/requirements.txt
requests
cryptography
django
python-dotenv
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-allauth
kiteconnect
langchain-google-genai
google-generativeai>=0.8.0
pytz
pinecone
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
pypdf>=3.17.0
yfinance
pandas
numpy
scikit-learn